Ethos-OCR GUI Mockup for DB Setup - Visualization
Ethos-OCR GUI Diagram Showing Layout
This is a GUI mockup of what the main screen of Ethos-OCR should look like.  Important to note now is the Visualization section at the bottom of the screen, the Database fields and type of data that should be stored from an OCR'ed screenshot, and the functionality within the Visualization section.