Screenshot of a botswarm comment
This is an example of a comment during a botswarm.  They typically will all have foreign themed names but the writing style is what to pay attention to.  These are the kinds of posts that will be matched from the Ethos-Botswarm module.  We will load in multiple examples like this one into the Ethos-OCR database and AI from other modules will watch incoming chat and come up with a percentage match of writing style.  Once the Ethos-OCR module comes online, I will feed it dozens of examples and train it to recognize attacks.